[task_scheduler] Add more tracing to RBE CAS code and warning if it takes longer than usual
This outage [1] took a while to figure out, due to a lack of logs
(no errors or warnings were thrown). This aims to address that by
making such anomalous behavior be visible in the logs.

As of writing, a healthy mergeCASInputs takes place in <10s.
The anomalous behavior saw that take many minutes.

It also removes the (mostly useless) calls to metrics2.FuncTimer
as those do not show up in the cloud logs and are harder to use
than the tracing spans that are used in their place.

[1] https://groups.google.com/a/google.com/g/luci-outages/c/tvtxib74WAo/m/omAz0QBOAgAJ

Change-Id: I83eae14f05f76077807a5c444c66553a40e7a986
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/763357
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Auto-Submit: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>