[cabe] explicitly name the api version as v1 in proto packages
Short version:

- Rename the gRPC service's proto package from `cabe.proto` to `cabe.v1`
- make `cabeserver` (temporarily) able to answer gRPC requests to *either* the old `cabe.proto.Analysis/GetAnalysis` or the new `cabe.v1.Analysis/GetAnalysis` gRPC names with the exact same handler code.

Longer version:

Why: In hindsight, `cabe.proto` was not a good choice for a gRPC service's proto package name. First, of *course* it's proto, so that doesn't need to be included in the name. Second, it lacks a version identifier. Including a version identifier in the service name is a fundamental API best practice, critical for us to be able to deploy more substantial (e.g. breaking) API version updates in the future.

What:
- Changes the proto package name from `cabe.proto` to `cabe.v1`
- Regenerates the generated go code
- Adds a hack to temporarily make cabeserver backwards compatible with existing client code.
- Adds a README.md in //cabe/proto to describe cabe's API versioning strategy.

This is the first of two CLs required to complete this "non-breaking mini API version roll". The other CL is on the catapult repo side, but it can land separately after this is deployed.

The order of CLs landing and production pushes should be:
- Land this CL and wait for it to deploy. It is designed to be backwards compatible with the grpc client currently in use by sandwich workflows in prod, so that it can be deployed by itself first, without causing issues for sandwich workflows.
- Verify that sandwich workflows' cloud functions (watch the logs for https://us-central1-chromeperf.cloudfunctions.net/get-cabe-analysis-prod in particular) can still make calls to `cabe.proto.Analysis` without issues at the grpc or proto level.  If not, roll this CL back  from prod. If calls from sandwich workflows to cabe.skia.org look okay though, proceed to the next step.
- Land https://chromium-review.googlesource.com/c/catapult/+/4963030 on the catapult side, and deploy it to production. This step is not as time sensitive since the old sandwich code can keep using the old proto package name until we update cabe to not answer to the old name any  more.
- If the above step works out, then at our convenience we'll write and deploy a cleanup CL to remove `alsoRegisterOldServiceProtoPackageName`

While most of the changes in this CL are just renaming things from `cabe.proto.*` to `cabe.v1.*`, there is one substantial new function the cabeserver main.go file: `alsoRegisterOldServiceProtoPackageName`. This is admittedly a hack, and includes comments for me to remove it once the catapult repo changes have made it to prod, so it should not live long.  It registers the same cpb.AnalyisService instance we're using with the new name, a second time using the old name. The second registration uses a modified `grpc.ServiceDesc` that allows it to also answer RPCs made to the old `cabe.proto` service name.

This "Answer requests made to both the old and the new service names with the exact same cpb.AnalysisServer instance" hack will only work in our particular case here because the wire formats of the Request and Response proto messages are identical between the two "services" - only the package names are different. Package names are not normally included in the serialized messages themselves, unless you use some specific proto extensions (which cabe doesn't use). I don't advise trying to do it this way in _any_ other circumstances besides this particular renaming use case.



Change-Id: Ie08480a4eb23c8ca7f897072331022b2c0b8ef73
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/769416
Reviewed-by: Joe Gregorio <jcgregorio@google.com>