[task scheduler] More logging and possible fix for split tryjob flow
My theory is that we're immediately and incorrectly marking the builds
as failed, because job.Done() incorrectly returning true, and
jobFinished marks the build a success if the job has
JOB_STATUS_SUCCESS and failure otherwise. Then buildbucket rejects the
build-started notification because the build is already finished.

I still don't understand the concurrent update errors, but maybe that's
caused by mixed usage of the cache and the DB directly.

Also rename some cache methods, as it's now obvious that these are
clearer.

Bug: b/288158829
Change-Id: I47fba5aecd83e6f42b8db607b2217b18a7c4799c
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/785299
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Eric Boren <borenet@google.com>