Change dfbuilder to not assume that all commits exist.
Previous to this CL dfbuilder presumed that all returns
from the following functions would be exact:

```
    tracestore.QueryTraces(ctx context.Context, tileNumber types.TileNumber, q *query.Query) (types.TraceSet, error)

    tracestore.ReadTraces(ctx context.Context, tileNumber types.TileNumber, keys []string) (types.TraceSet, error)

    tracestore.ReadTracesForCommitRange(ctx context.Context, keys []string, begin types.CommitNumber, end types.CommitNumber) (types.TraceSet, error)
```

That is, the TraceSet returned from QueryTraces and ReadTraces would be exactly the length of a Tile.

Similary ReadTracesForCommitRange presumed that every commit from `begin` to `end` would appear in the results, which is no longer the case.

The solution is to add a `[]provider.Commit` to each of those functions returns so that the called can map each returned value to the specific CommitNumber it corresponds to.

This CL started with a revert of "Calculate the tile commit range based on the data in Commits table.", and then added in the values to the above return signatures.

The revert will save 2 DB queries per call, and the data returned is still truncated, since data will only be returned for commits that exist.


Original change's description:
> Calculate the tile commit range based on the data in Commits table.
>
> Change-Id: I3492b6e7fbfcec3b96850e768a8bcddba6425368
> Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/675360
> Reviewed-by: Joe Gregorio <jcgregorio@google.com>
> Commit-Queue: Alex Sun <sunpeng@google.com>

Bugs: b/314806149
Change-Id: Ie1cc7dd5229e5a997a5996c193a4c7e4321a028a
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/789436
Reviewed-by: Alex Sun <sunpeng@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>