Unify chromeperf api clients.
Currently we have multiple clients for chromeperf apis which leads to a lot of code duplication. This change brings them together and adds some structure to make it cleaner and easier to maintain.
- Have a single ChromeperfClient interface/impl to send the http requests and parse response.
- Each api gets it's own interface/impl where it uses the ChromePerfClient to send the requests and add api specific semantics.
- All chromeperf clients are moved to a single package instead of being scattered around to promote reusability.

Most of the code here is structural (i.e moving things around). The functional aspects are unchanged for the most part.

Change-Id: I1d2151a7f62a3a77333aa7678e0abe3eb291429e
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/796525
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Sean McCullough <seanmccullough@google.com>
Commit-Queue: Ashwin Verleker <ashwinpv@google.com>