[task scheduler] Add more error context, cancel double-started jobs
We're seeing errors where we've called StartBuild multiple times for the
same Job. This shouldn't happen, because after the first successful
StartBuild we should have updated the Job in the DB with a new update
token and status JOB_STATUS_IN_PROGRESS. If we do see the job again in
startJob, we should see the it has already started and exit early.
Hopefully the additional error context helps diagnose the issue, and
canceling the job should prevent an infinite loop of StartBuild
attempts.

Bug: b/288158829
Change-Id: I82a9ee9d07641744bf00d25892762ee23b399418
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/801756
Reviewed-by: Ravi Mistry <rmistry@google.com>
Commit-Queue: Eric Boren <borenet@google.com>