[bisect] implement data aggregation in read_values
Data aggregation is referred to as "statistic" in Pinpoint.

This CL also adds gonum's stats package for standard deviation.

Regression detection workflow will send data aggregation to Pinpoint
to ensure the correct measurement. Without it, regressions may not
reproduce.

For example, here is a job without aggregation:
https://pinpoint-dot-chromeperf.appspot.com/job/1684cb07ae0000
Values are between 0 and 10 and no differences are found.

However, with aggregation, a difference can be repro'd:
https://pinpoint-dot-chromeperf.appspot.com/job/13d491a7ae0000

Bug: b/321311707
Change-Id: I2da22e5994cf51481f6f24a1c057634593be92aa
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/804277
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Commit-Queue: Leina Sun <sunxiaodi@google.com>