Fix edge case in newMergeWorker that caused TestBuildNew to be flaky.
If the commit is missing then req.commitNumberToOutputIndex returns 0, so the first value of the response trace gets overwritten.

Fix is to add explicit test for map membership (ok).

Made the test flaky because it was a race condition on which of the two tiles was handled by the merge worker first. If the second tile was handled second, then
the first value in the resulting trace would get overwritten with the wrong value.